requirements.txt
numpy==1.26.4
opencv-python-headless==4.8.1.78
paddlepaddle==2.6.1
paddleocr==2.7.3
paho-mqtt==1.6.1
Flask==3.0.3
PyYAML==6.0.2
requests==2.32.3
